refactor(dashboard): extract duplicated Observatory utils to shared module
entryTimestampMs (4 copies), hourToMs (2 copies), and isToolCall
(2 copies) were independently defined in observatory.ts, event-track,
tool-call-track, cross-signal-readout, and metric-track.

Consolidate all 8 definitions into observatory-utils.ts. Net -34
lines. Prevents future drift between copies.
